new evidence INSERT and SELECT tested and working, not implemented
